fix(krav): extract sync helpers out of 'use server' file to fix Turbopack build
Vercel preview build failed on feat/ai-tools-skills with:
  Server Actions must be async functions
  (at law-list-item-requirements.ts:101 resolveEffectiveAssignee,
   and law-list-item-requirements.ts:149 pickContentActionName)

Next.js 16 Turbopack enforces that every export from a 'use server'
file is an async function — a rule that pnpm typecheck (tsc --noEmit)
doesn't know about. Story 20.1 added two pure synchronous helpers to
the kravpunkter server-actions module; Story 20.1's QA-fix pass added
a third. All three tripped the rule at build time.

Fix: move the three helpers plus their types to a new non-'use server'
module at lib/requirements/helpers.ts. Action files import from the
new location; the server-actions module re-exports only the
EffectiveAssignee TYPE (type re-exports are compile-time only, no
runtime binding, so they remain legal inside 'use server').

No behavioural change — the helpers and their tests are unchanged;
only the module boundary moved.

Files
-----
- lib/requirements/helpers.ts (NEW) — home of resolveEffectiveAssignee
  + pickContentActionName + their types (EffectiveAssignee,
  ContentUpdatePatch, ContentUpdateExisting, RequirementContentActionName).
- app/actions/law-list-item-requirements.ts — removed the helper
  definitions + types; imports from the new lib; re-exports
  `type EffectiveAssignee` for any callers that still reference the
  old path.
- app/actions/workspace-requirements.ts — import redirected to
  @/lib/requirements/helpers.
- app/actions/__tests__/law-list-item-requirements.test.ts — import
  of resolveEffectiveAssignee + pickContentActionName redirected.
- tests/unit/app/actions/workspace-requirements.test.ts — import
  of resolveEffectiveAssignee redirected.

Validation
----------
- pnpm build: ✓ EXIT=0 (previously failed with Turbopack error)
- pnpm typecheck: ✓ 0 errors
- pnpm vitest run tests/unit app/actions/__tests__: ✓ 3763/3763
- Zero behavioural test changes — only import path churn
